[refactor/#20] LoginButton 파일 분리, 버튼 위치 조정
